# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/members/{id}": {
      "get": {
        "operationId": "AdminMemberController_getMemberDetail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get member detail (membership info + images)",
        "tags": ["Admin - Members"]
      },
      "delete": {
        "operationId": "AdminMemberController_deleteMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete member (deletedAt)",
        "tags": ["Admin - Members"]
      }
    },
...
    "/api/v1/members/{id}/restore": {
      "patch": {
        "operationId": "AdminMemberController_restoreMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore soft-deleted member",
        "tags": ["Admin - Members"]
      }
    },
    "/api/v1/membership-packages": {
      "get": {
        "operationId": "MembershipPackageController_findAll_v1",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all active/inactive membership packages",
        "tags": ["Admin - Membership Packages"]
      },
      "post": {
        "operationId": "MembershipPackageController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipPackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new membership package",
        "tags": ["Admin - Membership Packages"]
      }
    },
    "/api/v1/membership-packages/{id}": {
      "get": {
        "operationId": "MembershipPackageController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Membership Package ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get single package details",
        "tags": ["Admin - Membership Packages"]
      },
      "patch": {
        "operationId": "MembershipPackageController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Membership Package ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipPackageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update membership package",
        "tags": ["Admin - Membership Packages"]
      },
      "delete": {
        "operationId": "MembershipPackageController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Membership Package ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete package (deletedAt)",
        "tags": ["Admin - Membership Packages"]
      }
    },
    "/api/v1/membership-packages/{id}/status": {
      "patch": {
        "operationId": "MembershipPackageController_updateStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Membership Package ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackageStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate / deactivate package (isActive)",
        "tags": ["Admin - Membership Packages"]
      }
    },
    "/api/v1/membership-packages/{id}/restore": {
      "patch": {
        "operationId": "MembershipPackageController_restore_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Membership Package ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore soft-deleted package",
        "tags": ["Admin - Membership Packages"]
      }
    },
...
  },
  "info": {
    "title": "Gym Mini Check-in System API",
    "description": "API documentation for Gym Mini Check-in System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMembershipPackageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Package title",
            "example": "Monthly Membership"
          },
          "description": {
            "type": "string",
            "description": "Package description",
            "example": "Access to all gym facilities for one month"
          },
          "price": {
            "type": "number",
            "description": "Package price",
            "example": 50000
          },
          "durationDays": {
            "type": "number",
            "description": "Duration in days",
            "example": 30
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order for display",
            "example": 1,
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether package is active",
            "example": true,
            "default": true
          }
        },
        "required": ["title", "price", "durationDays"]
      },
      "UpdateMembershipPackageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Package title",
            "example": "Monthly Membership"
          },
          "description": {
            "type": "string",
            "description": "Package description",
            "example": "Access to all gym facilities for one month"
          },
          "price": {
            "type": "number",
            "description": "Package price",
            "example": 50000
          },
          "durationDays": {
            "type": "number",
            "description": "Duration in days",
            "example": 30
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order for display",
            "example": 1
          }
        }
      },
      "UpdatePackageStatusDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "description": "Package active status",
            "example": true
          }
        },
        "required": ["isActive"]
      },
...
    }
  }
}